Upgrade to HSL HFP API v2
Allows to listen to specific events instead of just
vehicle position updates. Now listening to door
open & close events.
